New hook: useWeb3.js and set listeners for accountsChanged, error and chainChanged
