+++ .env
@@ -1,1 +1,2 @@
 MONGODB_URI=mongodb+srv://admin-marketplace:PEDv6bJUZ2v3B1mJ@cluster0.k3pyg.mongodb.net/designers?retryWrites=true&w=majority&appName=Cluster0
+NODE_ENV=development
