+++ src/App.js
@@ -1,3 +1,4 @@
+import { useEffect } from "react";
 import { BrowserRouter } from "react-router-dom";
 import AboutUs from "./pages/AboutUs";
 import NavBar from "./components/navbar/Navbar";
@@ -8,8 +9,11 @@ import Loading from "./pages/Header";
 import Partners from "./pages/Partners";
 import Properties from "./pages/Properties";
 import Subscribe from "./pages/Subscribe";
+import useWeb3 from "./hooks/useWeb3";
 
 function App() {
+  const { setWalletListeners } = useWeb3();
+  useEffect(setWalletListeners, []);
   return (
     <>
       <BrowserRouter>
